+++ examples/stop.rs
@@ -4,7 +4,7 @@ use std::any::{Any, TypeId};
 use std::sync::Arc;
 use std::thread::sleep;
 use std::time::Duration;
-use tyractorsaur::prelude::{ActorRefTrait, ActorSystem, ActorTrait, Handler, MessageTrait, TyractorsaurConfig, StopMessage};
+use tyractorsaur::prelude::{ActorRefTrait, ActorSystem, ActorTrait, Handler, MessageTrait, TyractorsaurConfig, StopMessage, Context};
 
 #[derive(Clone)]
 struct TestMsg {}
@@ -25,9 +25,11 @@ impl ActorTrait for StopActor {
 }
 
 impl Handler<TestMsg> for StopActor {
-    fn handle(&mut self, msg: TestMsg) {
+    fn handle(&mut self, msg: TestMsg, context: &Context<Self>) {
+        context.actor_ref.send(TestMsg{});
         println!("Message received!");
         sleep(Duration::from_millis(100));
+        context.actor_ref.stop();
     }
 }
 
@@ -42,7 +44,7 @@ fn main() {
         .build(hw);
     // this is obviously handled, because it's the actor is still running
     x.send(TestMsg{});
-    x.stop();
+    //x.stop();
     // this is still handled, because the actor has not handled the stop Message yet
     x.send(TestMsg{});
     sleep(Duration::from_millis(200));