+++ src/context.rs
@@ -1,1 +1,12 @@
-pub struct Context {}
+use crate::prelude::ActorRef;
+use crate::actor::ActorTrait;
+use crate::system::ActorSystem;
+
+#[derive(Clone)]
+pub struct Context<A>
+where
+    A: ActorTrait + 'static
+{
+    pub actor_ref: ActorRef<A>,
+    pub system: ActorSystem,
+}